requirements.txt
# Core Dependencies
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1
redis==5.0.1
hiredis==2.2.3

# Network Communication
pyzmq==25.1.2
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# Data Structures & Hashing
mmh3==4.0.1
sortedcontainers==2.4.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.20.0
faker==22.0.0

# Monitoring & Metrics
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
colorama==0.4.6

# Security
cryptography==41.0.7
pyjwt==2.8.0

# Data Visualization (for reports)
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.26.2

# Documentation
pydantic==2.5.3
fastapi==0.108.0
uvicorn==0.25.0
